refactor: make separate image components
