+++ client/src/components/Card/TeamCard.js
@@ -1,9 +1,35 @@
 import React from 'react'
+import client from '../../sanity-client'
+import { useNextSanityImage } from 'next-sanity-image'
+import Gravatar from 'react-gravatar'
 
 import SocialMediaLinks from '../../sections/about/SocialMediaLinks'
 import { formatPhoneNumber } from '../../utils/helpers'
 import { Anchor, Typography } from '../Core'
 import { BaseCard } from './BaseCard'
+
+const TeamCardImage = ({ img, email, customImgAlt }) => {
+  const imageProps = useNextSanityImage(client, img?.asset._ref)
+  return (
+    <div className="tw-flex tw-justify-center">
+      <div className="tw-mt-6 tw-max-h-[160px] tw-min-h-[160px] tw-min-w-[160px] tw-max-w-[160px] tw-overflow-hidden tw-rounded-full">
+        {img ? (
+          <img
+            {...imageProps}
+            alt={customImgAlt ? customImgAlt : img.alt}
+            className="tw-order-1 tw-h-full tw-w-full tw-object-cover"
+          />
+        ) : (
+          <Gravatar
+            email={email}
+            className="tw-order-1 tw-h-auto tw-max-w-full"
+            size={200}
+          />
+        )}
+      </div>
+    </div>
+  )
+}
 
 export const TeamCard = ({
   link,
@@ -20,11 +46,12 @@ export const TeamCard = ({
   return (
     <BaseCard
       link={link}
-      img={img}
+      img={
+        <TeamCardImage img={img} email={email} customImgAlt={customImgAlt} />
+      }
       preTitle=""
       title={title}
       readMore={false}
-      customImgAlt={customImgAlt}
       gravatarEmail={email}
       centerContent={true}
     >
